+++ intelmq/bots/parsers/autoshun/parser.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 import html
 import html.parser
-import sys
 
 from intelmq.lib import utils
 from intelmq.lib.bot import ParserBot
@@ -19,14 +18,6 @@ TAXONOMY = {
 class AutoshunParserBot(ParserBot):
 
     def parse(self, report):
-        if sys.version_info[:2] == (3, 4):
-            # See https://docs.python.org/3/whatsnew/3.4.html#html
-            # https://docs.python.org/3/whatsnew/3.5.html#changes-in-the-python-api
-            # raises DeprecationWarning otherwise on 3.4, True by default in 3.5
-            self.parser = html.parser.HTMLParser(convert_charrefs=True)
-        else:
-            self.parser = html.parser.HTMLParser()
-
         raw_report = utils.base64_decode(report.get("raw"))
         splitted = raw_report.split("</tr>")
         self.tempdata = ['</tr>'.join(splitted[:2])]
@@ -43,10 +34,7 @@ class AutoshunParserBot(ParserBot):
 
         ip = info[1].split('</td>')[0].strip()
         last_seen = info[2].split('</td>')[0].strip() + '-05:00'
-        if sys.version_info < (3, 4):
-            description = self.parser.unescape(info[3].split('</td>')[0].strip())
-        else:
-            description = html.unescape(info[3].split('</td>')[0].strip())
+        description = html.parser.HTMLParser().unescape(info[3].split('</td>')[0].strip())
 
         for key in ClassificationType.allowed_values:
             if description.lower().find(key.lower()) > -1: